Renamed user_engine.rs to user_engine_api.rs
To reduce confusion with the similarly-named actor module. They are
related, but the API is an extension of the actor.
